added proto comments
