+++ ydb/mvp/meta/support_links/events.h
@@ -1,23 +1,9 @@
 #pragma once
 
+#include "types.h"
+
 #include <ydb/library/actors/core/event_local.h>
 #include <ydb/library/actors/core/events.h>
-
-namespace NMVP::NSupportLinks {
-
-struct TResolvedLink {
-    TString Title;
-    TString Url;
-};
-
-struct TSupportError {
-    TString Source;
-    TMaybe<ui32> Status;
-    TString Reason;
-    TString Message;
-};
-
-} // namespace NMVP::NSupportLinks
 
 namespace NMVP::NSupportLinks::TEvPrivate {
