+++ ydb/mvp/meta/ya.make
@@ -7,14 +7,6 @@ LIBRARY()
 SRCS(
     meta.cpp
     meta_cache.cpp
-    meta_cloud.h
-    meta_cluster.h
-    meta_clusters.h
-    meta_cp_databases.h
-    meta_cp_databases_verbose.h
-    meta_settings.h
-    meta_support_links.h
-    meta_db_clusters.h
     meta_versions.cpp
     mvp.cpp
 )
